requirements.txt
pyYaml
nextcord
colorama